Add synchronization with zigbeeHAgw (from zigbee-cc2530-serial).
Every time the driver starts, we wait until /var/run/zigbee.stable exists
before attempting to connect to the network gateway.

The objective is to prevent driver-go-zigbee generating a lot of noise
while zstack is starting (which can take a while and isn't completely reliable).

Of course, the correct fix is to make sure zstack starts reliably, but this is harder to achieve in the short term.

Signed-off-by: Jon Seymour <jon@ninjablocks.com>